Fix retroactive migrations
As the code and the schema fall out of sync, it's often possible
that there is a new attribute name in the "translates" array of
the model, but the translation column isn't yet in the database.

In practice, this breaks retroactive migrations if there are any
after_initialize callbacks trying to assign to said names, for
example.